Render and keymatch suggestions (#1)
* key: use byte literals / key values in switch

Use byte literals and vaxis.Key values in switch statement for clarity.

Correct matching of 'G'. Vaxis never sends an upper case ascii character
as the codepoint. Matching against this should use `Key.matches` or not
use a switch statement

* image: manage state to prevent reloading image each frame

Manage state of last selected path to prevent loading an image each
frame. Move the image.draw method outside of the conditional so it is
drawn each frame if we have one.